Update RGFW.h
On linux, when you drag and drop file and try to read the file path, sometimes the file path does not have a null terminating character - or at least it has junk after the file path. 

This fixed it for my use case, but I've only tested it in my app.